requirements.txt
dnspython~=2.6.1
Flask~=3.0.3
Flask-WTF~=1.2.1
pytest~=8.3.1
pytest-cov==5.0.0